requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0
httpx>=0.27.0,<0.29.0
opencv-python>=4.10.0,<4.13.0
celery[redis]>=5.4.0,<5.5.0
redis>=5.2.0,<6.0.0
pydantic-settings>=2.6.0,<2.8.0
python-multipart>=0.0.12,<0.1.0
tenacity>=9.0.0,<10.0.0
